fix bug in stat, move lstat64 back to unix
